+++ src/main/java/Duke.java
@@ -1,11 +1,9 @@
-import java.awt.print.PrinterAbortException;
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.PrintWriter;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
+
+import duke.command.TaskList;
+import duke.task.Deadlines;
+import duke.task.Events;
+import duke.task.ToDo;
+
 import java.util.Scanner;
 
 public class Duke {
@@ -84,499 +82,3 @@ public class Duke {
     }
 }
 
-abstract class Task {
-    String task_name;
-    String message_add;
-    String message_marked;
-    String message_unmarked;
-    String message_display;
-    String message_delete;
-    boolean done;
-    Parser parser;
-
-    Task(String name, boolean done) {
-        this.task_name = name;
-        this.done = done;
-        this.message_add = "";
-        this.message_marked = "";
-        this.message_unmarked = "";
-        this.message_delete = "";
-        parser = new Parser();
-    }
-
-    abstract void add();
-
-    abstract void marked();
-
-    abstract void unmarked();
-
-    abstract void display();
-
-    abstract void delete();
-}
-
-class ToDo extends Task {
-    final String raw;
-
-    ToDo(String name, boolean done) {
-        super(name, done);
-        raw = name;
-    }
-
-    @Override
-    void add() {
-        message_add = Parser.add + Parser.todoUnmarked_spaced + task_name;
-    }
-
-    @Override
-    void display() {
-        if (done) {
-            message_display = Parser.todoMarked + task_name;
-        } else {
-            message_display = Parser.todoUnmarked + task_name;
-        }
-
-    }
-
-    @Override
-    void delete() {
-        if (done)
-            message_delete = Parser.delete + Parser.todoMarked_spaced + task_name;
-        else
-            message_delete = Parser.delete + Parser.todoUnmarked_spaced + task_name;
-    }
-
-    @Override
-    void marked() {
-        message_marked = Parser.mark + Parser.todoMarked_spaced + task_name;
-        done = true;
-    }
-
-    @Override
-    void unmarked() {
-        message_unmarked = Parser.unmark + Parser.todoUnmarked_spaced + task_name;
-        done = false;
-    }
-
-}
-
-class Deadlines extends Task {
-
-    final String raw;
-    String endDate;
-    DateTranslator dateTranslator;
-
-    Deadlines(String name, boolean done) {
-        super(name, done);
-        raw = name;
-        //dateTranslator = new DateTranslator(raw);
-        extract();
-    }
-
-    void extract() {
-        try {
-            String[] tokens = task_name.split("/");
-            task_name = tokens[0];
-
-            if (!DateTranslator.is_date(raw)) {
-
-                String[] date = tokens[1].split(" ");
-                StringBuilder temp = new StringBuilder("(" + date[0] + ": ");
-                if (date.length > 3) {
-                    for (int x = 1; x < date.length; x++) {
-                        temp.append(date[x]).append(" ");
-                    }
-                    temp.append(")");
-                    endDate = temp.toString();
-                } else {
-                    endDate = date.length == 2 ? "(" + date[0] + ": " + date[1] + ")"
-                            : "(" + date[0] + ": " + date[1] + " " + date[2] + ")";
-                }
-
-            } else {
-
-                dateTranslator = new DateTranslator(raw);
-                endDate = dateTranslator.output;
-                System.out.println(endDate);
-            }
-        } catch (ArrayIndexOutOfBoundsException e) {
-            //exception forwarded to Task manager;
-        }
-
-    }
-
-    @Override
-    void add() {
-        message_add = Parser.add + Parser.deadlineUnmarked_spaced + task_name + endDate;
-    }
-
-    @Override
-    void display() {
-        if (done)
-            message_display = Parser.deadlinedoMarked + task_name + endDate;
-        else
-            message_display = Parser.deadlineUnmarked + task_name + endDate;
-    }
-
-    @Override
-    void delete() {
-        if (done)
-            message_delete = Parser.delete + Parser.deadlineMarked_spaced + task_name + endDate;
-        else
-            message_delete = Parser.delete + Parser.deadlineUnmarked_spaced + task_name + endDate;
-    }
-
-    @Override
-    void marked() {
-        message_marked = Parser.mark + Parser.deadlineMarked_spaced + task_name + endDate;
-        done = true;
-    }
-
-    @Override
-    void unmarked() {
-        message_unmarked = Parser.unmark + Parser.deadlineUnmarked_spaced + task_name + endDate;
-        done = false;
-    }
-}
-
-class Events extends Task {
-
-    final String raw;
-    String start;
-    String end;
-
-    Events(String name, boolean done) {
-        super(name, done);
-        raw = name;
-        extract();
-    }
-
-    void extract() {
-        try {
-            String[] tokens = task_name.split("/");
-            String[] startdate = tokens[1].split(" ");
-            String[] enddate = tokens[2].split(" ");
-            task_name = tokens[0];
-            start = startdate.length == 3 ? "(" + startdate[0] + ": " + startdate[1] + " " + startdate[2] + " "
-                    : "(" + startdate[0] + ": " + startdate[1] + " " + startdate[2] + " " + startdate[3] + " ";
-            end = enddate[0] + ": " + enddate[1] + ")";
-        } catch (ArrayIndexOutOfBoundsException e) {
-            //exception forwarded to Task manager;
-        }
-    }
-
-
-    @Override
-    void add() {
-        message_add = Parser.add + Parser.eventMarked_spaced + task_name + start + end;
-    }
-
-    @Override
-    void display() {
-        if (done)
-            message_display = Parser.eventMarked + task_name + start + end;
-        else
-            message_display = Parser.eventUnmarked + task_name + start + end;
-    }
-
-    @Override
-    void delete() {
-        if (done)
-            message_delete = Parser.delete + Parser.eventMarked_spaced + task_name + start + end;
-        else
-            message_delete = Parser.delete + Parser.eventUnmarked_spaced + task_name + start + end;
-    }
-
-    @Override
-    void marked() {
-        message_marked = Parser.mark + Parser.eventMarked_spaced + task_name + start + end;
-        done = true;
-    }
-
-    @Override
-    void unmarked() {
-        message_unmarked = Parser.unmark + Parser.eventUnmarked_spaced + task_name + start + end;
-        done = false;
-    }
-}
-
-class TaskList {
-    ArrayList<Task> ListOfTasks;
-    Storage storage;
-
-    TaskList(String path) {
-        //default size
-        ListOfTasks = new ArrayList<>(100);
-        storage = new Storage(path);
-        init();
-    }
-
-    void add(Task input) {
-        try {
-            if (input.task_name.equals("") || input.task_name.equals(" ")) {
-                throw new DukeException("OOPS!!! The description of a todo cannot be empty.\n");
-            } else {
-                ListOfTasks.add(input);
-                input.add();
-                System.out.println(input.message_add + "\n Now you have " + ListOfTasks.size() + " tasks in the list");
-                if (input instanceof ToDo) {
-                    storage.write("T|" + input.done + "|" + ((ToDo) input).raw);
-                } else if (input instanceof Deadlines) {
-                    storage.write("D|" + input.done + "|" + ((Deadlines) input).raw);
-                } else if (input instanceof Events) {
-                    storage.write("E|" + input.done + "|" + ((Events) input).raw);
-                } else {
-                    System.out.println("Unspecific type");
-                }
-            }
-        } catch (DukeException e) {
-            if (input instanceof ToDo)
-                System.out.println("OOPS!!! The description of a todo cannot be empty.\n");
-            else if (input instanceof Deadlines)
-                System.out.println("OOPS!!! The description of a deadline cannot be empty.\n");
-            else
-                System.out.println("OOPS!!! The description of a event cannot be empty.\n");
-        }
-    }
-
-    void displayAll() {
-        System.out.println(Parser.displaylist);
-        for (int x = 0; x < ListOfTasks.size(); x++) {
-            ListOfTasks.get(x).display();
-            System.out.println(x + 1 + ". " + ListOfTasks.get(x).message_display);
-        }
-    }
-
-    void mark(int index) {
-        try {
-            Task temp = ListOfTasks.get(index);
-            temp.marked();
-            System.out.println(temp.message_marked);
-            storage.markAt(index);
-        } catch (IndexOutOfBoundsException e) {
-            System.out.println("Invalid Index");
-        }
-    }
-
-    void unmark(int index) {
-        try {
-            Task temp = ListOfTasks.get(index);
-            temp.unmarked();
-            System.out.println(temp.message_unmarked);
-            storage.unmarkAt(index);
-        } catch (IndexOutOfBoundsException e) {
-            System.out.println("Invalid Index");
-        }
-    }
-
-    void delete(int index) {
-        try {
-            Task temp = ListOfTasks.get(index);
-            temp.delete();
-            ListOfTasks.remove(index);
-            System.out.println(temp.message_delete + "\n Now you have " + ListOfTasks.size() + " tasks in the list");
-            storage.deteleAt(index);
-        } catch (IndexOutOfBoundsException e) {
-            System.out.println("Invalid Index");
-        }
-    }
-
-    void init() {
-        storage.read();
-        if (!storage.records.isEmpty()) {
-            for (int x = 0; x < storage.records.size(); x++) {
-                String[] tokens = storage.records.get(x).split("\\|");
-                boolean status = Boolean.parseBoolean(tokens[1]);
-                if (tokens[0].equals("T")) {
-                    ListOfTasks.add(new ToDo(tokens[2], status));
-                } else if (tokens[0].equals("D")) {
-                    ListOfTasks.add(new Deadlines(tokens[2], status));
-                } else {
-                    ListOfTasks.add(new Events(tokens[2], status));
-                }
-            }
-        }
-    }
-
-    void file_writeAll() {
-        storage.WriteAll();
-    }
-
-
-}
-
-class DukeException extends Exception {
-    DukeException(String errorMessage) {
-        super(errorMessage);
-    }
-}
-
-class Storage {
-    FileWriter writer;
-    Scanner reader;
-    ArrayList<String> records;
-    String path;
-
-    Storage(String DES) {
-        try {
-            writer = new FileWriter(DES, true);
-            File store = new File(DES);
-            reader = new Scanner(store);
-            path = DES;
-            records = new ArrayList<>();
-        } catch (IOException e) {
-            System.out.println("Invalid file Path\n");
-        }
-    }
-
-    //write line by line
-    void write(String input) {
-        //writer.write(input);
-        records.add(input);
-        //writer.close();
-    }
-
-    void markAt(int index) {
-        String str = records.get(index).replace("false", "true");
-        records.set(index, str);
-    }
-
-    void unmarkAt(int index) {
-        String str = records.get(index).replace("true", "false");
-        records.set(index, str);
-    }
-
-    void deteleAt(int index) {
-        records.remove(index);
-    }
-
-    //read all
-    void read() {
-        while (reader.hasNextLine()) {
-            String data = reader.nextLine();
-            records.add(data);
-            //System.out.println(data);
-        }
-        reader.close();
-
-    }
-
-    void WriteAll() {
-        try {
-            clearFile();
-            for (String record : records) {
-                System.out.println(record);
-                writer.write(record);
-                writer.write(System.lineSeparator());
-            }
-            writer.flush();
-            writer.close();
-
-        } catch (IOException e) {
-            System.out.println("Write Error");
-        }
-    }
-
-    //transfer from records to File
-    void clearFile() throws IOException {
-        FileWriter f1 = new FileWriter("/Users/s.f/ip/src/Data/duke.txt", false);
-        PrintWriter p1 = new PrintWriter(f1, false);
-        p1.flush();
-        p1.close();
-        p1.close();
-    }
-
-}
-
-class DateTranslator {
-    String raw_input;
-    LocalDateTime time;
-    String output;
-
-    DateTranslator(String raw_input) {
-        this.raw_input = raw_input;
-        convert();
-    }
-
-    static boolean is_date(String raw_input) {
-        String[] token1 = raw_input.split("/");
-        String[] token2 = token1[1].split("-");
-        //first date  format
-        return token1.length > 2 || token2.length == 3;
-    }
-
-    void convert() {
-        //example input = return book /by 2/12/2019 1800
-        String[] tokens;
-        tokens = raw_input.split("/");
-        // 2/12/2019 1800 format
-        if (tokens.length > 2) {
-            //return book |by 2|12| 2019 1800
-            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-            int day = Integer.parseInt(tokens[1].split(" ")[1]); //remove by
-            int month = Integer.parseInt(tokens[2]);
-            String[] year_time = tokens[3].split(" ");
-            int year = Integer.parseInt(year_time[0]);
-            int time1;
-            int hour = 0;
-            int min = 0;
-            if (year_time.length > 1) {
-                time1 = Integer.parseInt(year_time[1]);
-                hour = time1 / 100;
-                min = time1 % 100;
-                formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss a");
-
-            }
-            time = LocalDateTime.of(year, month, day, hour, min);
-            output = time.format(formatter);
-            System.out.println(output);
-        } else {
-            //2019-12-02 1800 format
-            String[] Date_time = tokens[1].split(" ");
-            String[] date = Date_time[1].split("-");
-            int year;
-            int month;
-            int day;
-            int timing;
-            int hour = 0;
-            int min = 0;
-            year = Integer.parseInt(date[0]);
-            month = Integer.parseInt(date[1]);
-            day = Integer.parseInt(date[2]);
-            if (Date_time.length > 2) {
-                timing = Integer.parseInt(Date_time[2]);
-                hour = timing / 100;
-                min = timing % 100;
-            }
-            time = LocalDateTime.of(year, month, day, hour, min);
-            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MMM dd yyyy HH:mm a");
-            output = time.format(formatter);
-        }
-    }
-
-
-}
-
-class Parser {
-    static final String add = "Got it. I've added this task:\n";
-    static final String mark = "Nice! I've marked this task as done:\n";
-    static final String unmark = "OK, I've marked this task as not done yet:\n";
-    static final String displaylist = "Here are the tasks in your list\n";
-    static final String delete = "Noted. I've removed this task:\n";
-    static final String todoMarked = "[T][X]";
-    static final String todoUnmarked = "[T][ ]";
-    static final String todoMarked_spaced = "  [T][X] " ;
-    static final String todoUnmarked_spaced = " [T][ ] ";
-
-    static final String deadlinedoMarked = "[D][X] ";
-    static final String deadlineUnmarked = "[D][ ] ";
-    static final String deadlineMarked_spaced = "  [D][X] ";
-    static final String deadlineUnmarked_spaced = " [D][ ] ";
-
-    static final String eventMarked = "[E][X] ";
-    static final String eventUnmarked = "[E][ ] ";
-    static final String eventMarked_spaced = "  [E][X] ";
-    static final String eventUnmarked_spaced = " [E][ ] ";
-
-
-}
