+++ src/main/java/duke/utilities/Parser.java
@@ -1,0 +1,25 @@
+package duke.utilities;
+
+public class Parser {
+    public static final String add = "Got it. I've added this task:\n";
+    public static final String mark = "Nice! I've marked this task as done:\n";
+    public static final String unmark = "OK, I've marked this task as not done yet:\n";
+    public static final String displaylist = "Here are the tasks in your list\n";
+    public static final String delete = "Noted. I've removed this task:\n";
+    public static final String todoMarked = "[T][X]";
+    public static final String todoUnmarked = "[T][ ]";
+    public static final String todoMarked_spaced = "  [T][X] ";
+    public static final String todoUnmarked_spaced = " [T][ ] ";
+
+    public static final String deadlinedoMarked = "[D][X] ";
+    public static final String deadlineUnmarked = "[D][ ] ";
+    public static final String deadlineMarked_spaced = "  [D][X] ";
+    public static final String deadlineUnmarked_spaced = " [D][ ] ";
+
+    public static final String eventMarked = "[E][X] ";
+    public static final String eventUnmarked = "[E][ ] ";
+    public static final String eventMarked_spaced = "  [E][X] ";
+    public static final String eventUnmarked_spaced = " [E][ ] ";
+
+
+}
